+++ app/src/main/java/com/example/gads2020/ui/main_activity/MainActivityViewModelFactory.java
@@ -1,10 +1,11 @@
-package com.example.gads2020;
+package com.example.gads2020.ui.main_activity;
 
 import androidx.annotation.NonNull;
 import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 
 import com.example.gads2020.repo.LeadersRepo;
+import com.example.gads2020.ui.main_activity.MainActivityViewModel;
 
 public class MainActivityViewModelFactory implements ViewModelProvider.Factory {
     private LeadersRepo mRepo;
